requirements.txt
cx_oracle
django
bcrypt
djangorestframework
oracledb
django-cors-headers